third_party/blink/web_tests/platform/generic/external/wpt/scroll-animations/css/progress-based-animation-animation-longhand-properties.tentative-expected.txt

This is a testharness.js-based test.
FAIL animation-duration assert_equals: expected "25px" but got "none"
FAIL animation-duration: 0s assert_equals: expected "100px" but got "none"
FAIL animation-duration: infinite assert_equals: expected "100px" but got "none"
FAIL animation-iteration-count assert_equals: expected "25px" but got "none"
FAIL animation-iteration-count: 0 assert_equals: expected "0px" but got "none"
FAIL animation-iteration-count: infinite assert_equals: expected "100px" but got "none"
FAIL animation-direction: normal assert_equals: expected "25px" but got "none"
FAIL animation-direction: reverse assert_equals: expected "75px" but got "none"
FAIL animation-direction: alternate assert_equals: expected "20px" but got "none"
FAIL animation-direction: alternate-reverse assert_equals: expected "80px" but got "none"
FAIL animation-delay with a positive value assert_equals: expected "25px" but got "none"
FAIL animation-delay with a negative value assert_equals: expected "60px" but got "none"
FAIL animation-fill-mode assert_equals: expected "0px" but got "none"
Harness: the test ran to completion.
